+++ utils.h
@@ -1,2 +1,4 @@
+#include <Arduino.h>
+
 // format a float number to a given precision
-char *toPrecision(char *buffer, long len, float num, int precision);
+char *toPrecision(char *buffer, long len, float num, byte precision);
